# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WebAuthn/Passkey Authentication API with HttpOnly JWT cookies",
        "title": "Go Auth Core API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/passkeys": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Returns all passkeys registered by the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "passkeys"
                ],
                "summary": "List User Passkeys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.PasskeyResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/passkeys/{id}": {
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Deletes a passkey. Users can only delete their own passkeys and must keep at least one active.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "passkeys"
                ],
                "summary": "Delete Passkey",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Passkey ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Cannot delete last passkey",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Not authorized to delete this passkey",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Passkey not found",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Updates the name of a passkey. Users can only rename their own passkeys.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "passkeys"
                ],
                "summary": "Rename Passkey",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Passkey ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New Name",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RenamePasskeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Not authorized to modify this passkey",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Passkey not found",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.ErrorResponse": {
            "description": "Error details",
            "type": "object",
            "properties": {
                "code": {
                    "description": "Optional error code (e.g., TOKEN_EXPIRED)",
                    "type": "string",
                    "example": "TOKEN_EXPIRED"
                },
                "error": {
                    "description": "Error message\nexample: Invalid request format",
                    "type": "string",
                    "example": "Invalid request format"
                }
            }
        },
...
        "dtos.MessageResponse": {
            "description": "Generic response message",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Response message\nexample: Operation successful",
                    "type": "string",
                    "example": "Operation successful"
                }
            }
        },
...
        "handlers.PasskeyResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "last_used": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "MacBook Pro"
                }
            }
        },
        "handlers.RenamePasskeyRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "iPhone 15 Pro"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "JWT token in HttpOnly cookie (set automatically on login)",
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}